move some cases to correspond test file
- move test_cloud_init_config_ipv6, test_cloud_init_lineoverwrite
  test_cloud_init_userdata to test_cloud_init.py
- move test_kdump_unknown_nmi_panic_disabled,
  test_kdump_unknown_nmi_panic_enabled,test_hibernate_resume,
to test_lifecycle.py
- move test_network_device_hotplug to test_network_test.py
- delete test_vm_operation.py
- updated test_ping_arp_ping to run on aws
